% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize.R
\name{anonymize_dates}
\alias{anonymize_dates}
\title{Replace or remove dates and times from a character string}
\usage{
anonymize_dates(text, replace = FALSE)
}
\arguments{
\item{text}{text to remove dates and times from}

\item{replace}{logical Should dates and times be replaced with a generic token?}
}
\value{
text without dates and times
}
\description{
Replace or remove dates and times from a character string
}
\examples{
anonymize_dates("last friday, I took my friends fishing. That was the 13th of august. I'll never forget it, my first child was born the next year.", TRUE)

# it's not perfect...
anonymize_dates("2021-12-31, that's the day the year ends. We usually celebrate around midnight, or 12PM if you will - or if you're European, 00:00", TRUE)
}
