% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize.R
\name{anonymize_unknown_persons}
\alias{anonymize_unknown_persons}
\title{Removes or replaces all detected references to named persons in text}
\usage{
anonymize_unknown_persons(text, replace = FALSE)
}
\arguments{
\item{text}{text to anonymize}

\item{replace}{if TRUE, replace reference with generic tokens. If FALSE (the
default), removes references completely.}
}
\value{
anonymized text
}
\description{
Removes or replaces all detected references to named persons in text
}
\examples{

anonymize_unknown_persons("Hi, my name is John Doe. My favourite superhero is Deadpool. My dog, Fido, loves watching Spongebob Squarepants")
}
