% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.R
\name{extract_complement}
\alias{extract_complement}
\title{Obtain the relative complement of string a in string b}
\usage{
extract_complement(a, b)
}
\arguments{
\item{a}{text to obtain new content from}

\item{b}{previous version of text b}
}
\description{
Uses libdiff to determine a set of changes necessary to obtain a from b,
and returns new content only present in a.
}
\details{
Can be used to extract new content from an iterative set of documents, for
example logbooks created over time.
}
