% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_text}
\alias{extract_text}
\title{Extract text from a formatted document}
\usage{
extract_text(path, page = -1L)
}
\arguments{
\item{path}{\link{character} path to the document}

\item{page}{\link{numeric} for pdf documents, extracts the given page only. Will
attempt to automatically convert to pdf for other formats.}
}
\description{
Extracts text from a variety of popular document formats, including docx, pdf,
powerpoint, excel, open office, etc. Uses the \code{tika}, \code{pymupdf} and \code{docx2pdf}
python libraries to obtain a reasonably uniform output.
}
