% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize.R
\name{get_person_entities}
\alias{get_person_entities}
\title{Extract references to named persons from a text}
\usage{
get_person_entities(text)
}
\arguments{
\item{text}{}
}
\value{
list of named entities recognized as persons
}
\description{
Uses the default english model (en_core_web_md) available in spacy to extract
named entities.
}
