% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize.R
\name{anonymize_patterns}
\alias{anonymize_patterns}
\title{Remove or replace regular expression patterns}
\usage{
anonymize_patterns(text, patterns, replace = FALSE)
}
\arguments{
\item{text}{text to anonymize}

\item{patterns}{list of regular expressions to match}

\item{replace}{either a boolean value to remove references (FALSE), replace
references with a generic token (TRUE), or a list of replacement strings of the same
length as \code{patterns}.}
}
\value{
anonymized text
}
\description{
Remove or replace regular expression patterns
}
